open-sourcing/Backend/requirements.txt
Flask>=2.0.2,<3.0.0
Werkzeug>=2.0.2,<3.0.0
requests>=2.26.0,<3.0.0
python-dotenv==0.19.1
